\name{A112}
\docType{data}
\alias{A112}
\title{Copy Number Variant intensity data (from CNVtools)}
\description{
This data set has been obtained from CNVtools package in order to illustrate how CNVassoc and CNVtools compare
}
\usage{data(A112)}
%\format{A dataframe containing a row per individual (2593 in total) and 35 columns. The first two columns contain subject name and cohort membership.
%The remaining columns contain the copy number intensities for each locus within the CNV. This CNV contains 33 SNPs.}
\source{Obtained from CNVtools package (Wellcome Trust Case Control Consortium) }
\references{
\url{http://www.wtccc.org.uk/}
 "Genome-wide association study of 14,000 cases of seven common diseases and 3,000 shared controls."
 Wellcome Trust Case Control Consortium
 Nature. 2007;447;661-78. 
}
\keyword{datasets}
