\name{NeveCalled}
\alias{NeveCalled}
\title{Breast Cancer aCGH data with CGHcall}
\usage{data(NeveData)}
\description{
Breast cancer aCGH data called with CGHcall with default settings, containing 
2621 features and 50 samples
}
\format{
    An object of class CGHcall
}
\source{
Neve RM, Chin K, Fridlyand J et al. A collection of breast cancer cell lines 
for the study of functionally distinct cancer subtypes. \emph{Cancer Cell}, 
10:515-527, 2006.
Gonzalez JR, Subirana I, Escaramis G, Peraza S, Caceres A, Estivill X and 
Armengol L. Accounting for uncertainty when assessing association between copy 
number and disease: a latent class model. \emph{BMC Bioinformatics}, 
2009;10:172.
}
\value{
A list of two components: data frame with 54 variables and 2621 rows with the 
genotype data, and a vector containing the phenotype data.
}
\keyword{datasets}
