\name{NeveRegions}
\alias{NeveRegions}
\title{Breast Cancer aCGH data with CGHcall}
\usage{data(NeveRegions)}
\description{
Breast cancer aCGH data called with CGHcall with default settings, containing 2621 features and 50 samples
}

\format{
  An object of class CGHregions
}

\source{

Neve RM, Chin K, Fridlyand J et al. A collection of breast cancer cell lines for the study of functionally distinct cancer subtypes. \emph{Cancer Cell}, 10:515-527, 2006.

Gonzalez JR, Subirana I, Escaramis G, Peraza S, Caceres A, Estivill X and Armengol L. Accounting for uncertainty when assessing association between copy number and disease: a latent class model. \emph{BMC Bioinformatics}, 2009;10:172.

}

\keyword{datasets}

