\name{SNPTEST}
\alias{SNPTEST}
\alias{cases}
\alias{controls}
\title{Case-control data with SNPTEST format}
\usage{data(SNPTEST)}
\description{

Case-control data for 200 individuals (100 cases and 100 controls) and 50 SNPs using SNPTEST format (e.g. probabilities of AA, AB and BB genotypes)

}

\format{
Two objects: cases and controls. Each object is a matrix with 100 rows and 1505 columns. Column 1-5 correspond to annotation. Files 6-1505 correspond to the probabilities of being AA, AB or BB for 500 SNPs
}

\source{

 \url{https://mathgen.stats.ox.ac.uk/genetics_software/snptest/snptest.html}

}

\keyword{datasets}

