\name{SNPTEST}
\alias{SNPTEST}
\alias{cases}
\alias{controls}
\title{Case-control data with SNPTEST format}
\usage{data(SNPTEST)}
\description{

Case-control data for 200 individuals and 500 SNPs using SNPTEST format (e.g. probabilities of AA, AB and BB genotypes

}

\format{
Two objects: cases and controls. Each object is a matrix with 100 rows and 1505 columns. Column 1-5 correspond to annotation. Files 6-1505 correspond to the probabilities of being AA, AB or BB for 500 SNPs
}

\source{

 \url{http://www.stats.ox.ac.uk/~marchini/software/gwas/snptest.html}

}

\keyword{datasets}

